Promote verdict-side types to public; relocate renderers and adapters
A modularisation-prep cleanup. The motivating principle: a type that
appears in a public method signature on a public class must itself
be in a public package. Without this, a JPMS-modularised consumer
of punit could not compile against the verdict API without reaching
into framework internals — defeating the public/internal split.

PROMOTIONS (internal → public, principled relocations)

- TokenMode (enum, was nested inside
  internal.engine.budget.CostBudgetMonitor) →
  org.javai.punit.verdict.TokenMode, top-level. The enum is a
  verdict-side label naming how token cost was measured (NONE /
  STATIC / DYNAMIC); it is a peer to ThresholdOrigin and Verdict,
  not a sub-concept of the budget monitor. Exposed via
  ProbabilisticTestVerdict.CostSummary.tokenMode.

- PacingConfiguration (record, was internal.engine.pacing) →
  org.javai.punit.api.PacingConfiguration, top-level alongside the
  existing Pacing builder. The record carries authoring-time
  pacing constraints + computed execution plan; it is consumed
  by builder API and read off the verdict. Both user-facing.

The empty internal.engine.pacing/ package is removed.

RELOCATIONS (mislocated internal-shaped classes moved out of
public packages — same-direction moves, no API change)

- verdict.VerdictTextRenderer → internal.reporting.VerdictTextRenderer
- verdict.VerdictAdapter → internal.runtime.VerdictAdapter
- verdict.VerdictSinkBus → internal.reporting.VerdictSinkBus

The VerdictSink interface stays public at verdict.VerdictSink;
only the dispatcher (VerdictSinkBus) and the auxiliary renderer
(VerdictTextRenderer) and result adapter (VerdictAdapter) moved.

FREEZE-STORE RESIDUE: 116 → 52

Before the PR the publicMustNotImportInternal ArchUnit rule's
freeze store carried 116 entries across multiple categories.
After this PR it carries 52 — every remaining entry is intra-
module access from public packages into internal packages, which
JPMS export visibility does NOT restrict (cross-module is the
gated concern). Remaining residue:

- runtime.PUnit* (29) — facade pattern by design
- verdict.ProbabilisticTestVerdictBuilder (16) — private field
  types referencing internal engine types; not in public method
  signatures, so JPMS-compatible
- statistics.VerdictInterpreter (5) — pre-existing drift, uses
  internal.reporting.RateFormat; layering concern separate from
  JPMS, not in scope here
- api.covariate.CovariateDeclaration (2) — HashUtils intra-module
  use; private detail, JPMS-compatible

The PackageStructureArchitectureTest comment was updated to
reflect this.

NON-MOVES (out of scope)

VerdictInterpreter was considered for a parallel move to
internal.reporting/. That move was reverted because it produces
a new arrow that another architectural rule prohibits — junit5
ArchitectureTest's "Evaluators must not depend on reporting".
Resolving that drift is a separate layering cleanup (split
TestVerdictEvaluator's evaluation responsibility from its
rendering responsibility) and is tracked as follow-up.
