+++ docs/DEVELOPER-API.md
@@ -431,11 +431,11 @@ The packages an author imports and what each one carries.
 
 | Package                                                                               | Contains                                                                                                                                      | Author may import?                                            |
 |---------------------------------------------------------------------------------------|-----------------------------------------------------------------------------------------------------------------------------------------------|---------------------------------------------------------------|
-| `org.javai.punit.api`                                                                 | Annotations, UseCase, Contract, ContractBuilder, Sampling, Factor support, ValueMatcher, TestIntent, ThresholdOrigin, Pacing, TokenTracker, … | yes                                                           |
+| `org.javai.punit.api`                                                                 | Annotations, UseCase, Contract, ContractBuilder, Sampling, Factor support, ValueMatcher, TestIntent, ThresholdOrigin, Pacing, PacingConfiguration, TokenTracker, … | yes                                                           |
 | `org.javai.punit.api.spec`                                                            | Spec types (Spec, Experiment, ProbabilisticTest, Criterion, EvaluationContext, FactorsStepper, NextFactor, BaselineProvider, …)               | yes                                                           |
 | `org.javai.punit.api.covariate`                                                       | Covariate (interface) and built-in covariate categories                                                                                       | yes                                                           |
 | `org.javai.punit.runtime`                                                                      | `PUnit` entry point only — emitters live under `internal.runtime`                                                                             | yes — for `PUnit` only                                        |
-| `org.javai.punit.verdict`                                                                      | Verdict types, sinks, RunMetadata                                                                                                             | yes — for sink registration                                   |
+| `org.javai.punit.verdict`                                                                      | Verdict types, sinks, RunMetadata, `TokenMode` enum                                                                                            | yes — for sink registration                                   |
 | `org.javai.punit.statistics`                                                                   | Wilson, percentile, threshold derivation, feasibility evaluation                                                                              | yes — but rarely needed; the criteria already wrap statistics |
 | `org.javai.punit.internal.engine.*` (criteria, baseline, explore, optimize, covariate, spec, …) | Engine internals and concrete criterion impls                                                                                                 | **no** — internal                                             |
 | `org.javai.punit.internal.reporting`                                                           | Internal rendering helpers                                                                                                                    | **no** — internal                                             |
