+++ punit-core/src/main/java/org/javai/punit/api/PacingConfiguration.java
@@ -1,4 +1,4 @@
-package org.javai.punit.internal.engine.pacing;
+package org.javai.punit.api;
 
 import java.time.Instant;
