+++ punit-core/src/main/java/org/javai/punit/internal/runtime/VerdictAdapter.java
@@ -1,4 +1,4 @@
-package org.javai.punit.verdict;
+package org.javai.punit.internal.runtime;
 
 import java.util.List;
 
@@ -9,7 +9,11 @@ import org.javai.punit.api.spec.EngineRunSummary;
 import org.javai.punit.api.spec.EvaluatedCriterion;
 import org.javai.punit.api.spec.ProbabilisticTestResult;
 import org.javai.punit.api.spec.Verdict;
-import org.javai.punit.internal.engine.budget.CostBudgetMonitor.TokenMode;
+import org.javai.punit.verdict.PUnitVerdict;
+import org.javai.punit.verdict.ProbabilisticTestVerdict;
+import org.javai.punit.verdict.ProbabilisticTestVerdictBuilder;
+import org.javai.punit.verdict.RunMetadata;
+import org.javai.punit.verdict.TokenMode;
 import org.javai.punit.internal.engine.emit.LatencySection;
 import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.verdict.ProbabilisticTestVerdictBuilder.LatencyInput;
