+++ punit-core/src/test/java/org/javai/punit/internal/reporting/VerdictTextRendererTest.java
@@ -1,4 +1,4 @@
-package org.javai.punit.verdict;
+package org.javai.punit.internal.reporting;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,11 +8,14 @@ import java.util.Map;
 import java.util.Optional;
 
 import org.javai.punit.api.TestIntent;
-import org.javai.punit.internal.engine.budget.CostBudgetMonitor.TokenMode;
+import org.javai.punit.verdict.TokenMode;
 import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.api.UseCaseAttributes;
 import org.javai.punit.statistics.transparent.TransparentStatsConfig;
 import org.javai.punit.verdict.ProbabilisticTestVerdict.*;
+import org.javai.punit.verdict.PUnitVerdict;
+import org.javai.punit.verdict.ProbabilisticTestVerdict;
+import org.javai.punit.verdict.RunMetadata;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
