+++ punit-core/src/test/java/org/javai/punit/internal/runtime/VerdictAdapterTest.java
@@ -1,4 +1,4 @@
-package org.javai.punit.verdict;
+package org.javai.punit.internal.runtime;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -23,6 +23,10 @@ import org.javai.punit.api.spec.InconclusiveReasons;
 import org.javai.punit.api.spec.ProbabilisticTestResult;
 import org.javai.punit.api.spec.Verdict;
 import org.javai.punit.verdict.TerminationReason;
+import org.javai.punit.verdict.TokenMode;
+import org.javai.punit.verdict.PUnitVerdict;
+import org.javai.punit.verdict.ProbabilisticTestVerdict;
+import org.javai.punit.verdict.RunMetadata;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
