+++ punit-core/src/test/java/org/javai/punit/verdict/ProbabilisticTestVerdictBuilderTest.java
@@ -9,11 +9,11 @@ import java.util.Optional;
 import org.javai.punit.api.TestIntent;
 import org.javai.punit.api.ThresholdOrigin;
 import org.javai.punit.api.spec.Verdict;
-import org.javai.punit.internal.engine.budget.CostBudgetMonitor.TokenMode;
+import org.javai.punit.verdict.TokenMode;
 import org.javai.punit.internal.engine.budget.SharedBudgetMonitor;
 import org.javai.punit.internal.engine.budget.SharedBudgetMonitor.Scope;
 import org.javai.punit.api.BudgetExhaustedBehavior;
-import org.javai.punit.internal.engine.pacing.PacingConfiguration;
+import org.javai.punit.api.PacingConfiguration;
 import org.javai.punit.verdict.ExpirationPolicy;
 import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.internal.engine.spec.model.ExecutionSpecification;
