+++ punit-report/src/test/java/org/javai/punit/report/VerdictAdapterIntegrationTest.java
@@ -34,7 +34,7 @@ import org.javai.punit.api.spec.ProbabilisticTestResult;
 import org.javai.punit.api.spec.Verdict;
 import org.javai.punit.verdict.ProbabilisticTestVerdict;
 import org.javai.punit.verdict.RunMetadata;
-import org.javai.punit.verdict.VerdictAdapter;
+import org.javai.punit.internal.runtime.VerdictAdapter;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
