+++ punit-report/src/test/java/org/javai/punit/report/VerdictTextRendererTest.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import org.javai.punit.api.TestIntent;
-import org.javai.punit.internal.engine.budget.CostBudgetMonitor.TokenMode;
+import org.javai.punit.verdict.TokenMode;
 import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.api.UseCaseAttributes;
 import org.javai.punit.verdict.ProbabilisticTestVerdict;
@@ -23,7 +23,7 @@ import org.javai.punit.verdict.ProbabilisticTestVerdict.StatisticalAnalysis;
 import org.javai.punit.verdict.ProbabilisticTestVerdict.Termination;
 import org.javai.punit.verdict.ProbabilisticTestVerdict.TestIdentity;
 import org.javai.punit.verdict.PUnitVerdict;
-import org.javai.punit.verdict.VerdictTextRenderer;
+import org.javai.punit.internal.reporting.VerdictTextRenderer;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
