+++ punit-sentinel/src/test/java/org/javai/punit/sentinel/testsubjects/SentinelTestSubjects.java
@@ -8,12 +8,12 @@ import java.util.Optional;
 import org.javai.punit.api.Experiment;
 import org.javai.punit.api.ProbabilisticTest;
 import org.javai.punit.api.TestIntent;
-import org.javai.punit.internal.engine.budget.CostBudgetMonitor.TokenMode;
+import org.javai.punit.verdict.TokenMode;
 import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.api.UseCaseAttributes;
 import org.javai.punit.verdict.ProbabilisticTestVerdict;
 import org.javai.punit.verdict.PUnitVerdict;
-import org.javai.punit.verdict.VerdictSinkBus;
+import org.javai.punit.internal.reporting.VerdictSinkBus;
 import org.opentest4j.AssertionFailedError;
 
 /**
